feat: more error codes added

+++ backend/src/controllers/RequestController.js
@@ -38,9 +38,14 @@ requestController.getAll = async function (req, res) {
 };
 
 requestController.getOne = async function (req, res) {
-  const { id } = req.params;
-  const request = await Request.findById(id);
-  res.json({ request });
+  try {
+    const { id } = req.params;
+    const request = await Request.findById(id);
+    res.json({ request });
+  } catch (e) {
+    res.status(404);
+    res.json({ err: e });
+  }
 };
 
 requestController.create = async function (req, res) {
@@ -172,14 +177,22 @@ requestController.update = async function (req, res) {
     }
   } catch (e) {
     res.status(500);
-    res.json({});
+    res.json({ err: e });
   }
 };
 
 requestController.delete = async function (req, res) {
   const { id } = req.params;
-  await Request.findByIdAndDelete(id);
-  res.json({ delete: true });
+
+  const request = await Request.findById(id);
+
+  if (request) {
+    await Request.findByIdAndDelete(id);
+    res.json({ delete: true });
+  } else {
+    res.status(404);
+    res.json({ delete: false });
+  }
 };
 
 module.exports = requestController;
