+++ backend/src/controllers/UserController.js
@@ -15,7 +15,8 @@ userController.getAll = async function (req, res) {
     const userList = await User.find(filters);
     res.json({ userList });
   } catch (e) {
-    res.json({});
+    res.status(500);
+    res.json({ err: e });
   }
 };
 
@@ -26,6 +27,7 @@ userController.getOne = async function (req, res) {
     res.json({ user });
   } catch (e) {
     res.status(404);
+    res.json({ err: e });
   }
 };
 
@@ -49,20 +51,35 @@ userController.create = async function (req, res) {
     }
   } catch (e) {
     res.status(500);
-    res.json({});
+    res.json({ err: e });
   }
 };
 
 userController.update = async function (req, res) {
   const { id } = req.params;
-  const result = await User.findByIdAndUpdate(id, req.body);
-  res.json({ result });
+  const user = await User.findById(id);
+
+  if (user) {
+    const result = await User.findByIdAndUpdate(id, req.body);
+    res.json({ result });
+  } else {
+    res.status(404);
+    res.json({});
+  }
 };
 
 userController.delete = async function (req, res) {
   const { id } = req.params;
-  await User.findByIdAndDelete(id);
-  res.json({ delete: true });
+
+  const user = await User.findById(id);
+
+  if (user) {
+    await User.findByIdAndDelete(id);
+    res.json({ delete: true });
+  } else {
+    res.status(404);
+    res.json({ delete: false });
+  }
 };
 
 module.exports = userController;
